Очищение всего кода по стилю
