+++ src/main/java/gyber/org/mainCore/MainCoreApplication.java
@@ -1,9 +1,6 @@
 package gyber.org.mainCore;
 
-import gyber.org.mainCore.data.entities.Role;
-import gyber.org.mainCore.data.entities.User;
-import gyber.org.mainCore.data.service.UserService;
-import org.springframework.boot.CommandLineRunner;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
@@ -12,17 +9,21 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import java.util.ArrayList;
+
 
 @SpringBootApplication
-@EnableJpaRepositories(basePackages = "gyber.org.mainCore.data.repository")
-@EntityScan(basePackages = "gyber.org.mainCore.data.entities")
+@EnableJpaRepositories(basePackages = "gyber.org.mainCore.repository")
+@EntityScan(basePackages = "gyber.org.mainCore.entity")
 public class MainCoreApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(MainCoreApplication.class, args);
 	}
 
+	/**
+	 * Получение encoder пароля.
+	 * @return encoder пароля
+	 */
 	@Bean
 	PasswordEncoder passwordEncoder() {
 		return new BCryptPasswordEncoder();
