--- src/main/java/gyber/org/mainCore/controllers/MainController.java
@@ -1,18 +1,0 @@
-package gyber.org.mainCore.controllers;
-
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/demo-controller")
-public class MainController {
-
-    @GetMapping
-    public ResponseEntity<String> sayHello() {
-        return ResponseEntity.ok("Hello from secured endpoint");
-    }
-
-}
