--- src/main/java/gyber/org/mainCore/data/entities/enums/TypeMediaFile.java
@@ -1,10 +1,0 @@
-package gyber.org.mainCore.data.entities.enums;
-
-public enum  TypeMediaFile {
-
-    DOCUMENT ,
-    VIDEO ,
-    PHOTO ,
-    OTHER_FILE
-
-}
