--- src/main/java/gyber/org/mainCore/data/notifications/SystemNotification.java
@@ -1,15 +1,0 @@
-package gyber.org.mainCore.data.notifications;
-
-import gyber.org.mainCore.notifications.Notification;
-
-import java.util.Date;
-
-public class SystemNotification extends Notification {
-
-
-    public SystemNotification(String from, String assignedTo, Date dateOfSend, String message) {
-        super(from, assignedTo, dateOfSend, message, false);
-    }
-
-    public SystemNotification(){}
-}
