--- src/main/java/gyber/org/mainCore/data/repository/RoleRepository.java
@@ -1,9 +1,0 @@
-package gyber.org.mainCore.data.repository;
-
-import gyber.org.mainCore.data.entities.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, Long> {
-
-    Role findByName (String name);
-}
