--- src/main/java/gyber/org/mainCore/data/service/UserServiceImpl.java
@@ -1,77 +1,0 @@
-package gyber.org.mainCore.data.service;
-
-import gyber.org.mainCore.data.entities.Role;
-import gyber.org.mainCore.data.entities.User;
-import gyber.org.mainCore.data.repository.RoleRepository;
-import gyber.org.mainCore.data.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-@Slf4j
-public class UserServiceImpl implements UserService, UserDetailsService {
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByNickNameOrEmail(username, username);
-        if(user == null) {
-            log.error("User not found in the database");
-            throw new UsernameNotFoundException("User not found in the database");
-        } else {
-            log.info("User found in the database: ", username);
-        }
-        Collection<SimpleGrantedAuthority> authorities = new ArrayList<>();
-        user.getRoles().forEach(role -> authorities.add(new SimpleGrantedAuthority(role.getName())));
-        return new org.springframework.security.core.userdetails.User(user.getNickName(), user.getPasswd(), authorities);
-    }
-
-    @Override
-    public User saveUser(User user) {
-        log.info("Saving new user {} to the database", user.getNickName());
-        user.setPasswd(passwordEncoder.encode(user.getPasswd()));
-        return userRepository.save(user);
-    }
-
-    @Override
-    public Role saveRole(Role role) {
-        log.info("Saving new role {} to the database", role.getName());
-        return roleRepository.save(role);
-    }
-
-    @Override
-    public void addRoleToUser(String username, String roleName) {
-        log.info("Add new role {} to user {} to the database", roleName, username);
-        User user = userRepository.findByNickNameOrEmail(username, username);
-        Role role = roleRepository.findByName(roleName);
-        user.getRoles().add(role);
-    }
-
-    @Override
-    public User getUser(String username) {
-        log.info("Fetching user {}", username);
-        return userRepository.findByNickNameOrEmail(username, username);
-    }
-
-    @Override
-    public List<User> getUsers() {
-        log.info("Fetching all users");
-        return userRepository.findAll();
-    }
-
-}
