+++ src/main/java/gyber/org/mainCore/entity/enums/TypeMediaFile.java
@@ -1,0 +1,10 @@
+package gyber.org.mainCore.entity.enums;
+
+public enum TypeMediaFile {
+
+    DOCUMENT,
+    VIDEO,
+    PHOTO,
+    OTHER_FILE
+
+}
