+++ src/main/java/gyber/org/mainCore/notification/SystemNotification.java
@@ -1,0 +1,23 @@
+package gyber.org.mainCore.notification;
+
+import java.util.Date;
+
+public class SystemNotification extends Notification {
+
+    /**
+     * AllArgsConstructor.
+     * @param from ...
+     * @param assignedTo ...
+     * @param dateOfSend ...
+     * @param message ...
+     */
+    public SystemNotification(String from, String assignedTo, Date dateOfSend, String message) {
+        super(from, assignedTo, dateOfSend, message, false);
+    }
+
+    /**
+     * NoArgsConstructor.
+     */
+    public SystemNotification() {
+    }
+}
